exec abstraction with DockerCompose method
This was split out from
I8e3a6cd393edbdc7c83875d3775babb8480b30f2
which was generally agreed to be a good idea.

I still think having some sort of controlled name for the
docker-compose project is a good idea but will try to introduce
that in another change.

Change-Id: I65e38407ede88203c2a7f2d25cbd8e664e37f3de