Prefix docker-compose project name with mwcli
This patch used to switch the project name to include the
username that was running the project, but that ended up
not being a great solution and removed the ability for a
single user to run multiple environments.

This final version just prefixes the project name with mwcli
so that it is easier to understand in the context of other
docker-compose projects. This will appear as "mwcli-core" by default.

There is still potential to look at expanding this more in the
future, perhaps re introducing user in some way, perhaps in an
optional way. But this is a good start.

Change-Id: I8e3a6cd393edbdc7c83875d3775babb8480b30f2