Add 'docker status' command
This is split from the mwdd port patch.
I09a4cc905534fc2750b007ca0da31912009c8c8d

In that patch this was added as part of another command allowing
direct access to docker-compose.
We may still want that down the line, but review stated that it might
be nicer to include this as part of the 'docker' command for now,
and this is that patch.

The exact review conversation was at:
https://gerrit.wikimedia.org/r/c/mediawiki/tools/cli/+/597864/9/cmd/dc.go

"status" is used instead of "ps" to hide the docker layer and make
everything more user friendly.
The "ps" command that is run is still visable using the --verbose option

Bug: T253313
Change-Id: If38a28cdd64789530107d4e91de29341920a0029