Add --verbose flag to docker command
This came up in a few reviews, most recently in
If38a28cdd64789530107d4e91de29341920a0029 which needs
t be debugged.

--verbose will output the command that has been run, as well
as the stdout and stderr output from the command.

This alters the places that will have verbose output to stop
using combined stderr and stdout output.

IMO this is on the whole an improvement for now, but we might want
to cleanup this execution even further in the future.
It would also be nice to have this debug mode for ALL executed commands
not just the ones that I have selected under the docker command so far.

Example output can be found at https://phabricator.wikimedia.org/P12262

Change-Id: Ic248eb20023992ef7caa2cad04177bde0b4e18a9