Refactor error handling
Refactor so that output and errors when using exec.RunCommand are all
handled in handleCommandRun in order to provide general handling of
output while also handling specific cases.

Also prints stdout and any errors regardless of verbosity level because
it is very confusing to run a command and not have any indication
whether it succeeded or failed. Verbosity now only controls the printing
of the shell command that resulted from the mw cli command.

Change-Id: I7c5530ca9680466d1edecc531df07ce5af8452b8