Move spinner to HandlerOptions
This brings some consistency to the way that optional
execution parameters are handled.

In order to look right the spinner needs to be tied into the
execution code and error handling / verbose output code.
This is why it was pushed down into exec before HandlerOptions
existed.
Now HandlerOptions do exist this seems like the right place for
the spinner to live.

This was extracted from I7d1a66dd96b293d3fe3d6b371212e34018872a49

Change-Id: I7da770b4d819f75cb4cd8bac695c348487d06001