Refactor internal env code to a type
This is further prep in the direction of
I7d1a66dd96b293d3fe3d6b371212e34018872a49
that is useful in any case.

The internal/env package now exposes a DotFile type
which has the various methods attached for interacting
with a .env file.

The .env file is currently tied to the MediaWiki core directory
so the commands have a check added for the current location.
This is already done for the docker commands, but not here yet.

Use of the type should make using .env files in different locations
easier in the future.

Change-Id: I3bf2079ff4be2f8589d17443639d5e4e03dadf12