Refactor internal mediawiki code to a type
This is further prep in the direction of
I7d1a66dd96b293d3fe3d6b371212e34018872a49
that is useful in any case.

The internal/mediawiki package now exposes a MediaWiki type
which has the various methods attached for interacting
with a MediaWiki directory.

The internal/docker package contains functionality of the mediawiki-docker
dev environment.

This means there is now a clear split between the CLI / user / presentation
layer, which exists in the cmd package.
The buisness layer that lives in other packages.
internal/mediawiki: things generically performed to mediawiki on disk
internal/docker: things specifica to the mediawiki-docker command / environment

Change-Id: I487866628a87d11eeb4d467d5c718a706f7c6e23