Refactor docker env to allow use by other commands
This refactors the way the docker env commands are created
so that the same cobra commands can be used in other places
if needed.
This will be used for the mwdd command, which also wants to
allow people to edit the env vars via cli.

Reusing the commands ensures some consistency.

anon functions are used to pass the directory into the command
currently, as we want mediawikiOrFatal() to run when the command
is run, not when the commands are setup.

Change-Id: Ie5515c79b9d9d5578cde3f262166fb7069b8f50d