mwdd: Installing with sqlite
This is a feature from mediawiki-docker that makes more
sense now that mwdd can be flexible and have any number of
services running / not running at a given time.

This means that the minimum setup is now the base + mediawiki.
MySQL things will be added after.

Includes most of the LocalSettings stuff used in mwdd
(though it isn't all actively used here yet)

Includes addition of a custom special page to allow users
to easily see the status of some dev env related services
though not all of those services exist yet.

MySQL support will be added in a follow up patch.

This code was verified on github actions on addshores fork
https://github.com/addshore/mwcli/actions/runs/780347275
You can see the .github workflow there to see how to test this
(As docs have not been added to this repo yet)

Bug: T253313
Change-Id: I7d1a66dd96b293d3fe3d6b371212e34018872a49