Make sure env set can be used before the file exists
We wan't an `env set` command to create the .env file, and
directories on the way to the file if needed.
Otherwise users can try to set things before running any other
commands and it won't work.

I came across this while writing some CI test stuff and wanting
to set the dir of mediawiki.
I would set, then run `mwdd create`, but the set would not have
dont anything!

Change-Id: I4d6cd57ced68993e38e2e506a007bc34f2371585