Enable installing with sqlite
This is a feature from mediawiki-docker that makes more
sense now that mwdd can be flexible and have any number of
services running / not running at a given time.

This means that the minimum setup is now the base + mediawiki.
MySQL things will be added after.

This includes various fixes needed to make mediawiki "fully" work
such as file permissions, setting tweaks etc.

Change-Id: Ie7c30d45425b5df03bde2e2e8c2999febe1180a7