mwdd: Enable installing with mariadb
This adds the logic for the mysql service that for now
just uses a mariadb image from docker hub.
Eventually a followup will come to allow custom images to be used here.

Verified in github actions on addshore fork
https://github.com/addshore/mwcli/actions/runs/780426092
You can see the steps needed to test this out in the test there.

Assuming you have mediawiki, composer installed with the correct settings
file, then:
- mwcli mwdd create
- mwcli mwdd mysql create
- mwcli mwdd mediawiki install somesite mysql

Really should figure out how to best maintain the docs for this soon.
Vuepress perhaps?

Bug: T253313
Change-Id: I04b56065757e0fad21d85cd76b6624e7ea86ebde