README updates (new directories, dev.sh convenience script etc)
Various changes have been happening but the README is not totally up to date.
So update it!

And also add a ./dev.sh script which is now the recommended way to run the
tool in its un-built fashion.
This takes care of making sure that the static files are rebuild every time
you run the script, as this as tripped me over quite a few times during rapid
development.

Another alternative would be to force people to build each time, but IMO
we shouldn't force this as it just ends up being slower.

Change-Id: I6cbd916086c5177d5c0f98a03659e6d5d06ba467