mwdd: Implement nice destroy commands for individual services
These commands now destroy the containers and volumes of the
specific service being interacted with.
Prior to this the mediawiki destroy command would just run
down on the whole docker-compose setup (as it was the only service).
But now we need to be a little more selective.

docker-compose doesn't provide commands to do all of this out of
the box, which is why a docker command is needed to destroy specific
named volumes for the services.

Bug: T253313
Change-Id: I8a861978894dd0fe10ffbcadc73314d64138ddd3