mwdd: Initial phpunit command
This can be tested with a command such as:

./dev.sh mwdd mediawiki phpunit \
/var/www/html/w/tests/phpunit/unit/includes/FormOptionsTest.php

This uses the docker SDK directly to allow for easy passthrough
of stdin and out for commands.
This similar pattern will be used for further CLI interaction
in mwdd.

The phpunit command will also be worked on in further patches, including:
 - Clever things to do with directories (and maybe path completion?)
 - Ability to run on different --wikis

Change-Id: I4ca04705ba584e3eb241bd75e33e3530305fbf82