mwdd: mw phpunit & composer, use "correct" working dir
This means that when these commands are used on the host
system in a directory such as "skins/Vector", then working
dir used with docker will reflect this.

This allows composer commands to be used on sub repos, like
skins and extensions, as well as for phpunit tests to easily
be run when in other directories with auto completion of paths
that just works.

This does NOT allow users to use these commands when outside of
the expected MediaWiki directory.
This is out of scope for this patch (but might be something to
consider in the future)

Bug: T282350
Change-Id: Id5fb148116bb3346f42867af2140933619f9bf2f