Output more useful version info
This will make it easier for people to give feedback while
using various development builds
(This will also be useful when we do real releases)

This required a bunch of changes to the MakeFile to
actually get the build dependencies loading when required.
The way "staticfiles" was done in the past wasn't great.

Change-Id: Icdbe959dcf429305ebcec57cbe64bd05b0ded737