Add dev alias, replace docker with mwdd
Creates an alias called dev for the default development environment.
This will be useful when other environments are added to the cli.
Replaces the docker command with mwdd to remove duplicate functionality.

Also update the test to be mwdd / docker focused
but this needs to be commented out for now...

Change-Id: Ibb2798f438f9372289694f0bb4ce5dbb087d18dc