Stop printing success update notes twice
If the update was a success, then that update message is also printed in
cmd/version.go

In general the code in internal should not print to the console.
However, since execution stops there if the update was not a success,
the output is moved into that if-condition to keep showing the error
message to the user before the program exits.