mwdd: Fix sqlite (and other file issues) with umods & chmods
This alters the location of the sqlite db files.
It also sets a umod ALWAYS in install.php and also via
LocalSettings.php which means we should have far less
permission issues when juggling differnet things in different
contexts.

It might make snense to set umods more too, for things such as
running composer update as the wrong user etc..

https://phabricator.wikimedia.org/T287654