Add half baked permission checks to update command
It's hard to check the writability of the currently running
executable, so for now just check the writability of the dir
which seems to improve this situation a little?
Especially as the issue reported in
https://phabricator.wikimedia.org/T293963
is actually to do with the /usr/bin/.mw.new file, which I guess
is the result of the download?

Perhaps in the future we should just stop using this update
library all together and write our own?