Make dev exec command work with docker-compose v2
A change between v1 and v2 of docker-compose meant that
container names moves from using _s to -s for their parts.

mwcli depends on this name riht now and thus breaks with v2.
This change removes the dep on names, instead selecting containers
based on labels which remain the same between docker-compose version.

Thanks for @migr for spotting the issue.