util/jsonfile: Add clear method and some tests
I was going to use this while tracking state of which services
are running as part of the setup, but it turns out it will be
far easier to just use docker-compose ps to figure this out,
than tracking it in a json file.
But this small change and tests look good still!