Progress bar while update is being loaded
This is not ideal, as it isn't actually tied to
the progress of the download or saving of the file.
But it is better than just a blank and confusing space.

In the future the update process would want a little more
refactoring to enable status to be reported back to the UI
layer.

Bug: T293586