Fix telemetry on "docker env" cmds & tests
the env command previously ignored the persistent
pre run, so as to avoid the mwdd wizard.
Instead we now just don't run the wizard for specific sub
command branches, so we can continue to use persistent pre run.

This refactoring should also enable fixing of T292331 which is
another case of the wizard running when we do not want it.

This commit also comes with tests for the cobra utils package.