Entirely rewrite config handeling
This makes use of koanf, which means we can now have
defaults, that are overridden by json config, which
can be overridden by env vars.

And in the future this will also tie into cli flags

Probably fixes other issues to do with config handling,
and will also reduce reads and writes on disk around the config.

Bug: T373052